frontend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.1
scikit-learn==1.3.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
deep-translator==1.11.4
joblib==1.3.2
gunicorn==21.2.0
numpy==1.26.0
